Flock

Behavior Monad

type Behavior a = ReaderT Plane (StateT Agent (Rand StdGen)) a

runBehavior :: Behavior a -> StdGen -> Plane -> Agent -> (a, Agent, StdGen)

Type

data Agent

Agents are circular autonomous mobile objects, searching of there destination.

Constructors

Agent

Fields

_agentPosition :: Point

Position of the agent on the plane

_agentRadius :: Float

Radius of the agent

_agentDirection :: Vector

Direction, in which the agent is heading

_agentSpeed :: Float

Speed at which the agent is moving

_agentSensorRange :: Distance

Sensor range of the agent

_agentFlocking :: Bool

Is the agent in a flock?

_agentAtDest :: Bool

Is the agent or its flock at its destination?

Constructor

mkAgent

Arguments

:: Point
Starting position

-> Float
Radius of the agent (size)

-> Vector
Initial direction

-> Float
Initial speed

-> Distance
Sensor range

-> Agent

Constructs an agent

Control functions

move :: Behavior ()

Moves an agent, according to its state;

scan :: Behavior ([Agent], [Obstacle])

Scans the plane around an agent, returning all agents and obstacles withing it's sensor range

turn :: Angle -> Behavior ()

Turns an agent by an angle

turnTowards :: Point -> Behavior ()

Lets the agent directly head to a specific point.

turnAway :: Point -> Behavior ()

Lenses

agentPosition :: Lens' Agent Point

agentRadius :: Lens' Agent Float

agentDirection :: Lens' Agent Vector

agentSpeed :: Lens' Agent Float

agentSensorRange :: Lens' Agent Distance

agentFlocking :: Lens' Agent Bool

agentAtDest :: Lens' Agent Bool

Obstacles

Type

data Obstacle

Obstacles are circular objects, which can not be passed by any agent.

Constructors

Obstacle

Fields

_obstPosition :: Point

The position of the obstacle on the plane

_obstRadius :: Float

The radius of the obstacle

Lenses

obstRadius :: Lens' Obstacle Float

obstPosition :: Lens' Obstacle Point

Plane

Type

data Plane

Constructors

Plane

Fields

_planeAgents :: [Agent]

_planeObstacles :: [Obstacle]

Control functions

step

Arguments

:: (Plane -> Agent -> Agent)
Behavior function.

-> Plane
Plane

-> Plane

Simulates the plane by one step according to a function, which defines the behavior of an agent in the plane.

stepR :: (Plane -> Agent -> Rand g Agent) -> Plane -> Rand g Plane

As step, but in a random monad.

Lenses

planeObstacles :: Lens' Plane [Obstacle]

planeAgents :: Lens' Plane [Agent]

Geometry

type Angle = Float

type Distance = Float

Rendering

class Render a where

Minimal complete definition

render

Methods

render :: Plane -> a -> Picture

DEBUG

Utils

dist :: Point -> Point -> Float

collidesWith :: (IsObject a, IsObject b) => a -> b -> Bool

Checks, whether two objects collide.

distance :: (IsObject a, IsObject b) => a -> b -> Distance

Calculates the edge-to-edge-distance between two objects.

angle :: IsObject a => Agent -> a -> Angle

Returns the Angle between the direction vector and the vector between the agent and another object;

ATTENTION! The angle returned is symmetrical for both sides; i.e.: if the angle is, e.g., PI/2, it is either 90 Degrees on the left or on the right side.

isOnLeft :: IsObject a => Agent -> a -> Bool

Checks, whether a give object is on the left side of the agent.

isOnRight :: IsObject a => Agent -> a -> Bool

Checks, whether a give object is on the right side of the agent.

scanD

Arguments

:: Angle
Direction of view

-> Angle
Field of view

-> Behavior ([Agent], [Obstacle])

Scans only in a certain direction, with a certain field of view

lookThere :: IsObject a => Angle -> Angle -> Agent -> a -> Bool

class IsObject obj where

A class of objects in the simulated plane.

Minimal complete definition

radius, position

Methods

radius :: obj -> Float

Returns the radius of an object.

position :: obj -> Point

Returns the position of an object.